pyproject.toml
[project]
name = "kafka-test"
version = "0.1.0"
description = ""
authors = [
    {name = "genii01",email = "genii0627@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "datasets (>=3.2.0,<4.0.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "transformers (>=4.48.1,<5.0.0)",
    "torch (>=2.5.1,<3.0.0)",
    "torchvision (>=0.20.1,<0.21.0)",
    "datetime (>=5.5,<6.0)",
    "efficientnet-pytorch (>=0.7.1,<0.8.0)",
    "isort (>=5.13.2,<6.0.0)",
    "mypy (>=1.14.1,<2.0.0)",
    "autoflake (>=2.3.1,<3.0.0)",
    "onnx (>=1.17.0,<2.0.0)",
    "onnxruntime (>=1.20.1,<2.0.0)",
    "fastapi (>=0.110.0,<1.0.0)",
    "uvicorn (>=0.27.1,<1.0.0)",
    "python-multipart (>=0.0.9,<1.0.0)",
    "pydantic (>=2.6.3,<3.0.0)",
    "pydantic-settings (>=2.7.1,<3.0.0)",
    "flake8 (>=7.1.1,<8.0.0)",
    "black (>=24.10.0,<25.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'
extend-exclude = '''
# A regex preceded with ^/ will apply only to files and directories
# in the root of the project.
^/docs
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
skip_gitignore = true

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true

[tool.poetry.group.dev.dependencies]
isort = "^5.12.0"
mypy = "^1.5.0"
autoflake = "^2.2.0"
